requirements.txt
openai
Pillow
transformers
torch
timm
einops
scipy
ftfy
easyocr
numpy
opencv-python

